using collection tests for map/set
