+++ types/map_test.go
@@ -2,10 +2,210 @@ package types
 
 import (
 	"bytes"
+	"math/rand"
+	"sort"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/suite"
 )
+
+const testMapSize = 1000
+
+type testMapGenFn func(v Number) Value
+type testMapLessFn func(x, y Value) bool
+
+type testMap struct {
+	entries     []mapEntry
+	less        testMapLessFn
+	tr          *Type
+	knownBadKey Value
+}
+
+func (tm testMap) Len() int {
+	return len(tm.entries)
+}
+
+func (tm testMap) Less(i, j int) bool {
+	return tm.less(tm.entries[i].key, tm.entries[j].key)
+}
+
+func (tm testMap) Swap(i, j int) {
+	tm.entries[i], tm.entries[j] = tm.entries[j], tm.entries[i]
+}
+
+func (tm testMap) SetValue(i int, v Value) testMap {
+	entries := make([]mapEntry, 0, len(tm.entries))
+	entries = append(entries, tm.entries...)
+	entries[i].value = v
+	return testMap{entries, tm.less, tm.tr, tm.knownBadKey}
+}
+
+func (tm testMap) Remove(from, to int) testMap {
+	entries := make([]mapEntry, 0, len(tm.entries)-(to-from))
+	entries = append(entries, tm.entries[:from]...)
+	entries = append(entries, tm.entries[to:]...)
+	return testMap{entries, tm.less, tm.tr, tm.knownBadKey}
+}
+
+func (tm testMap) toMap() Map {
+	keyvals := []Value{}
+	for _, entry := range tm.entries {
+		keyvals = append(keyvals, entry.key, entry.value)
+	}
+	return NewTypedMap(tm.tr, keyvals...)
+}
+
+func (tm testMap) toCompoundMap() compoundMap {
+	keyvals := []Value{}
+	for _, entry := range tm.entries {
+		keyvals = append(keyvals, entry.key, entry.value)
+	}
+	return NewTypedMap(tm.tr, keyvals...).(compoundMap)
+}
+
+func (tm testMap) Flatten(from, to int) []Value {
+	flat := make([]Value, 0, len(tm.entries)*2)
+	for _, entry := range tm.entries[from:to] {
+		flat = append(flat, entry.key)
+		flat = append(flat, entry.value)
+	}
+	return flat
+}
+
+func (tm testMap) FlattenAll() []Value {
+	return tm.Flatten(0, len(tm.entries))
+}
+
+func newTestMap(length int) testMap {
+	entries := make([]mapEntry, 0, length)
+	for i := 0; i < length; i++ {
+		entry := mapEntry{Number(i), Number(i * 2)}
+		entries = append(entries, entry)
+	}
+	return testMap{entries,
+		func(x, y Value) bool {
+			return !y.(OrderedValue).Less(x.(OrderedValue))
+		},
+		MakeMapType(NumberType, NumberType), Number(length + 2)}
+}
+
+func newTestMapWithGen(length int, gen testMapGenFn, less testMapLessFn, tr *Type) testMap {
+	s := rand.NewSource(4242)
+	used := map[int64]bool{}
+
+	var mask int64 = 0xffffff
+	entries := make([]mapEntry, 0, length)
+	for len(entries) < length {
+		v := s.Int63() & mask
+		if _, ok := used[v]; !ok {
+			entry := mapEntry{gen(Number(v)), gen(Number(v * 2))}
+			entries = append(entries, entry)
+			used[v] = true
+		}
+	}
+
+	return testMap{entries, less, MakeMapType(tr, tr), gen(Number(mask + 1))}
+}
+
+type mapTestSuite struct {
+	collectionTestSuite
+	elems testMap
+}
+
+func newMapTestSuite(size uint, expectRefStr string, expectChunkCount int, expectPrependChunkDiff int, expectAppendChunkDiff int) *mapTestSuite {
+	length := 1 << size
+	elems := newTestMap(length)
+	tr := MakeMapType(NumberType, NumberType)
+	tmap := NewTypedMap(tr, elems.FlattenAll()...)
+	return &mapTestSuite{
+		collectionTestSuite: collectionTestSuite{
+			col:                    tmap,
+			expectType:             tr,
+			expectLen:              uint64(length),
+			expectRef:              expectRefStr,
+			expectChunkCount:       expectChunkCount,
+			expectPrependChunkDiff: expectPrependChunkDiff,
+			expectAppendChunkDiff:  expectAppendChunkDiff,
+			validate: func(v2 Collection) bool {
+				l2 := v2.(Map)
+				out := []Value{}
+				l2.IterAll(func(key, value Value) {
+					out = append(out, key)
+					out = append(out, value)
+				})
+				return valueSlicesEqual(elems.FlattenAll(), out)
+			},
+			prependOne: func() Collection {
+				dup := make([]mapEntry, length+1)
+				dup[0] = mapEntry{Number(-1), Number(-2)}
+				copy(dup[1:], elems.entries)
+				flat := []Value{}
+				for _, entry := range dup {
+					flat = append(flat, entry.key)
+					flat = append(flat, entry.value)
+				}
+				return NewTypedMap(tr, flat...)
+			},
+			appendOne: func() Collection {
+				dup := make([]mapEntry, length+1)
+				copy(dup, elems.entries)
+				dup[len(dup)-1] = mapEntry{Number(length*2 + 1), Number((length*2 + 1) * 2)}
+				flat := []Value{}
+				for _, entry := range dup {
+					flat = append(flat, entry.key)
+					flat = append(flat, entry.value)
+				}
+				return NewTypedMap(tr, flat...)
+			},
+		},
+		elems: elems,
+	}
+}
+
+func TestMapSuite1K(t *testing.T) {
+	suite.Run(t, newMapTestSuite(10, "sha1-db0aa902943bcbe03520bb58f58a1d808285f350", 16, 2, 2))
+}
+
+func TestMapSuite4K(t *testing.T) {
+	suite.Run(t, newMapTestSuite(12, "sha1-c601acb4295a8260a8a8d479c6a26e6cef9281ef", 56, 2, 2))
+}
+
+func getTestNativeOrderMap(scale int) testMap {
+	return newTestMapWithGen(int(mapPattern)*scale, func(v Number) Value {
+		return v
+	}, func(x, y Value) bool {
+		return !y.(OrderedValue).Less(x.(OrderedValue))
+	}, NumberType)
+}
+
+func getTestRefValueOrderMap(scale int) testMap {
+	setType := MakeSetType(NumberType)
+	return newTestMapWithGen(int(mapPattern)*scale, func(v Number) Value {
+		return NewTypedSet(setType, v)
+	}, func(x, y Value) bool {
+		return !y.Ref().Less(x.Ref())
+	}, setType)
+}
+
+func getTestRefToNativeOrderMap(scale int, vw ValueWriter) testMap {
+	refType := MakeRefType(NumberType)
+	return newTestMapWithGen(int(mapPattern)*scale, func(v Number) Value {
+		return vw.WriteValue(v)
+	}, func(x, y Value) bool {
+		return !y.(Ref).TargetRef().Less(x.(Ref).TargetRef())
+	}, refType)
+}
+
+func getTestRefToValueOrderMap(scale int, vw ValueWriter) testMap {
+	setType := MakeSetType(NumberType)
+	refType := MakeRefType(setType)
+	return newTestMapWithGen(int(mapPattern)*scale, func(v Number) Value {
+		return vw.WriteValue(NewTypedSet(setType, v))
+	}, func(x, y Value) bool {
+		return !y.(Ref).TargetRef().Less(x.(Ref).TargetRef())
+	}, refType)
+}
 
 func TestNewMap(t *testing.T) {
 	assert := assert.New(t)
@@ -45,6 +245,31 @@ func TestMapUniqueKeysNumber(t *testing.T) {
 	assert.True(Number(5).Equals(m.Get(Number(1))))
 }
 
+func TestMapHas(t *testing.T) {
+	if testing.Short() {
+		t.Skip("Skipping test in short mode.")
+	}
+	assert := assert.New(t)
+
+	vs := NewTestValueStore()
+	doTest := func(tm testMap) {
+		m := tm.toMap()
+		m2 := vs.ReadValue(vs.WriteValue(m).TargetRef()).(compoundMap)
+		for _, entry := range tm.entries {
+			k, v := entry.key, entry.value
+			assert.True(m.Has(k))
+			assert.True(m.Get(k).Equals(v))
+			assert.True(m2.Has(k))
+			assert.True(m2.Get(k).Equals(v))
+		}
+	}
+
+	doTest(getTestNativeOrderMap(16))
+	doTest(getTestRefValueOrderMap(2))
+	doTest(getTestRefToNativeOrderMap(2, vs))
+	doTest(getTestRefToValueOrderMap(2, vs))
+}
+
 func TestMapHasRemove(t *testing.T) {
 	assert := assert.New(t)
 	m1 := NewMap()
@@ -56,6 +281,43 @@ func TestMapHasRemove(t *testing.T) {
 	assert.False(m1.Has(NewString("foo")))
 	assert.True(m2.Has(NewString("foo")))
 	assert.False(m3.Has(NewString("foo")))
+}
+
+func TestMapRemove(t *testing.T) {
+	if testing.Short() {
+		t.Skip("Skipping test in short mode.")
+	}
+	assert := assert.New(t)
+
+	doTest := func(incr int, tm testMap) {
+		whole := tm.toMap()
+		run := func(i int) {
+			expected := tm.Remove(i, i+1).toMap()
+			actual := whole.Remove(tm.entries[i].key)
+			assert.Equal(expected.Len(), actual.Len())
+			assert.True(expected.Equals(actual))
+		}
+		for i := 0; i < len(tm.entries); i += incr {
+			run(i)
+		}
+		run(len(tm.entries) - 1)
+	}
+
+	doTest(128, getTestNativeOrderMap(32))
+	doTest(64, getTestRefValueOrderMap(4))
+	doTest(64, getTestRefToNativeOrderMap(4, NewTestValueStore()))
+	doTest(64, getTestRefToValueOrderMap(4, NewTestValueStore()))
+}
+
+func TestMapRemoveNonexistentKey(t *testing.T) {
+	assert := assert.New(t)
+
+	tm := getTestNativeOrderMap(2)
+	original := tm.toMap()
+	actual := original.Remove(Number(-1)) // rand.Int63 returns non-negative numbers.
+
+	assert.Equal(original.Len(), actual.Len())
+	assert.True(original.Equals(actual))
 }
 
 func TestMapFirst(t *testing.T) {
@@ -79,6 +341,26 @@ func TestMapFirst(t *testing.T) {
 	assert.True(ev.Equals(av))
 }
 
+func TestMapFirst2(t *testing.T) {
+	if testing.Short() {
+		t.Skip("Skipping test in short mode.")
+	}
+	assert := assert.New(t)
+
+	doTest := func(tm testMap) {
+		m := tm.toMap()
+		sort.Stable(tm)
+		actualKey, actualValue := m.First()
+		assert.True(tm.entries[0].key.Equals(actualKey))
+		assert.True(tm.entries[0].value.Equals(actualValue))
+	}
+
+	doTest(getTestNativeOrderMap(16))
+	doTest(getTestRefValueOrderMap(2))
+	doTest(getTestRefToNativeOrderMap(2, NewTestValueStore()))
+	doTest(getTestRefToValueOrderMap(2, NewTestValueStore()))
+}
+
 func TestMapSetGet(t *testing.T) {
 	assert := assert.New(t)
 	m1 := NewMap()
@@ -95,6 +377,58 @@ func TestMapSetGet(t *testing.T) {
 	assert.True(Number(42).Equals(m2.Get(NewString("foo"))))
 	assert.True(Number(43).Equals(m3.Get(NewString("foo"))))
 	assert.Nil(m4.Get(NewString("foo")))
+}
+
+func TestMapSet(t *testing.T) {
+	if testing.Short() {
+		t.Skip("Skipping test in short mode.")
+	}
+	assert := assert.New(t)
+
+	doTest := func(incr, offset int, tm testMap) {
+		expected := tm.toMap()
+		run := func(from, to int) {
+			actual := tm.Remove(from, to).toMap().SetM(tm.Flatten(from, to)...)
+			assert.Equal(expected.Len(), actual.Len())
+			assert.True(expected.Equals(actual))
+		}
+		for i := 0; i < len(tm.entries)-offset; i += incr {
+			run(i, i+offset)
+		}
+		run(len(tm.entries)-offset, len(tm.entries))
+		assert.Panics(func() {
+			expected.Set(Number(1), Bool(true))
+		}, "Should panic due to wrong type")
+	}
+
+	doTest(18, 3, getTestNativeOrderMap(9))
+	doTest(128, 1, getTestNativeOrderMap(32))
+	doTest(64, 1, getTestRefValueOrderMap(4))
+	doTest(64, 1, getTestRefToNativeOrderMap(4, NewTestValueStore()))
+	doTest(64, 1, getTestRefToValueOrderMap(4, NewTestValueStore()))
+}
+
+func TestMapSetExistingKeyToNewValue(t *testing.T) {
+	if testing.Short() {
+		t.Skip("Skipping test in short mode.")
+	}
+	assert := assert.New(t)
+
+	tm := getTestNativeOrderMap(2)
+	original := tm.toMap()
+
+	expectedWorking := tm
+	actual := original
+	for i, entry := range tm.entries {
+		newValue := Number(int64(entry.value.(Number)) + 1)
+		expectedWorking = expectedWorking.SetValue(i, newValue)
+		actual = actual.Set(entry.key, newValue).(compoundMap)
+	}
+
+	expected := expectedWorking.toMap()
+	assert.Equal(expected.Len(), actual.Len())
+	assert.True(expected.Equals(actual))
+	assert.False(original.Equals(actual))
 }
 
 func TestMapSetM(t *testing.T) {
@@ -117,6 +451,30 @@ func TestMapDuplicateSet(t *testing.T) {
 	assert := assert.New(t)
 	m1 := NewMap(Bool(true), Bool(true), Number(42), Number(42), Number(42), Number(42))
 	assert.Equal(uint64(2), m1.Len())
+}
+
+func TestMapMaybeGet(t *testing.T) {
+	if testing.Short() {
+		t.Skip("Skipping test in short mode.")
+	}
+	assert := assert.New(t)
+
+	doTest := func(tm testMap) {
+		m := tm.toMap()
+		for _, entry := range tm.entries {
+			v, ok := m.MaybeGet(entry.key)
+			if assert.True(ok, "%v should have been in the map!", entry.key) {
+				assert.True(v.Equals(entry.value), "%v != %v", v, entry.value)
+			}
+		}
+		_, ok := m.MaybeGet(tm.knownBadKey)
+		assert.False(ok, "m should not contain %v", tm.knownBadKey)
+	}
+
+	doTest(getTestNativeOrderMap(2))
+	doTest(getTestRefValueOrderMap(2))
+	doTest(getTestRefToNativeOrderMap(2, NewTestValueStore()))
+	doTest(getTestRefToValueOrderMap(2, NewTestValueStore()))
 }
 
 func TestMapIter(t *testing.T) {
@@ -162,7 +520,59 @@ func TestMapIter(t *testing.T) {
 	assert.True(got(NewString("a"), Number(0)) || got(NewString("b"), Number(1)))
 }
 
-func TestMapFilter(t *testing.T) {
+func TestMapIter2(t *testing.T) {
+	assert := assert.New(t)
+
+	doTest := func(tm testMap) {
+		m := tm.toMap()
+		sort.Sort(tm)
+		idx := uint64(0)
+		endAt := uint64(mapPattern)
+
+		m.Iter(func(k, v Value) (done bool) {
+			assert.True(tm.entries[idx].key.Equals(k))
+			assert.True(tm.entries[idx].value.Equals(v))
+			if idx == endAt {
+				done = true
+			}
+			idx++
+			return
+		})
+
+		assert.Equal(endAt, idx-1)
+	}
+
+	doTest(getTestNativeOrderMap(16))
+	doTest(getTestRefValueOrderMap(2))
+	doTest(getTestRefToNativeOrderMap(2, NewTestValueStore()))
+	doTest(getTestRefToValueOrderMap(2, NewTestValueStore()))
+}
+
+func TestMapIterAll(t *testing.T) {
+	if testing.Short() {
+		t.Skip("Skipping test in short mode.")
+	}
+	assert := assert.New(t)
+
+	doTest := func(tm testMap) {
+		m := tm.toMap()
+		sort.Sort(tm)
+		idx := uint64(0)
+
+		m.IterAll(func(k, v Value) {
+			assert.True(tm.entries[idx].key.Equals(k))
+			assert.True(tm.entries[idx].value.Equals(v))
+			idx++
+		})
+	}
+
+	doTest(getTestNativeOrderMap(16))
+	doTest(getTestRefValueOrderMap(2))
+	doTest(getTestRefToNativeOrderMap(2, NewTestValueStore()))
+	doTest(getTestRefToValueOrderMap(2, NewTestValueStore()))
+}
+
+func TestMapFilter2(t *testing.T) {
 	assert := assert.New(t)
 
 	m := NewMap(Number(0), NewString("a"), Number(1), NewString("b"), Number(2), NewString("c"))
@@ -402,4 +812,113 @@ func TestMapChunks(t *testing.T) {
 	l3 := NewMap(Number(0), NewTypedRefFromValue(Number(1)))
 	c3 := l3.Chunks()
 	assert.Len(c3, 1)
+}
+
+func TestMapFilter(t *testing.T) {
+	if testing.Short() {
+		t.Skip("Skipping test in short mode.")
+	}
+	assert := assert.New(t)
+
+	doTest := func(tm testMap) {
+		m := tm.toMap()
+		sort.Sort(tm)
+		pivotPoint := 10
+		pivot := tm.entries[pivotPoint].key
+		actual := m.Filter(func(k, v Value) bool {
+			return tm.less(k, pivot)
+		})
+		assert.True(newTypedMap(tm.tr, tm.entries[:pivotPoint+1]...).Equals(actual))
+
+		idx := 0
+		actual.IterAll(func(k, v Value) {
+			assert.True(tm.entries[idx].key.Equals(k), "%v != %v", k, tm.entries[idx].key)
+			assert.True(tm.entries[idx].value.Equals(v), "%v != %v", v, tm.entries[idx].value)
+			idx++
+		})
+	}
+
+	doTest(getTestNativeOrderMap(16))
+	doTest(getTestRefValueOrderMap(2))
+	doTest(getTestRefToNativeOrderMap(2, NewTestValueStore()))
+	doTest(getTestRefToValueOrderMap(2, NewTestValueStore()))
+}
+
+func TestMapFirstNNumbers(t *testing.T) {
+	if testing.Short() {
+		t.Skip("Skipping test in short mode.")
+	}
+	assert := assert.New(t)
+
+	mapType := MakeMapType(NumberType, NumberType)
+
+	kvs := []Value{}
+	for i := 0; i < testMapSize; i++ {
+		kvs = append(kvs, Number(i), Number(i+1))
+	}
+
+	m := NewTypedMap(mapType, kvs...)
+	assert.Equal("sha1-4d2fb53d2020630376f2b188addf58e546dc33e8", m.Ref().String())
+	height := deriveMapHeight(m)
+	cm := m.(compoundMap)
+	assert.Equal(height, cm.tuples[0].childRef.Height())
+}
+
+func TestMapRefOfStructFirstNNumbers(t *testing.T) {
+	if testing.Short() {
+		t.Skip("Skipping test in short mode.")
+	}
+	assert := assert.New(t)
+	vs := NewTestValueStore()
+
+	structType := MakeStructType("num", TypeMap{
+		"n": NumberType,
+	})
+	refOfTypeStructType := MakeRefType(structType)
+
+	mapType := MakeMapType(refOfTypeStructType, refOfTypeStructType)
+
+	kvs := []Value{}
+	for i := 0; i < testMapSize; i++ {
+		k := vs.WriteValue(NewStruct(structType, structData{"n": Number(i)}))
+		v := vs.WriteValue(NewStruct(structType, structData{"n": Number(i + 1)}))
+		assert.NotNil(k)
+		assert.NotNil(v)
+		kvs = append(kvs, k, v)
+	}
+
+	m := NewTypedMap(mapType, kvs...)
+	assert.Equal("sha1-c179c3dd34ec4430b11515cb5f859b8f4546b970", m.Ref().String())
+	height := deriveMapHeight(m)
+	cm := m.(compoundMap)
+	// height + 1 because the leaves are Ref values (with height 1).
+	assert.Equal(height+1, cm.tuples[0].childRef.Height())
+}
+
+func TestMapModifyAfterRead(t *testing.T) {
+	assert := assert.New(t)
+	vs := NewTestValueStore()
+	m := getTestNativeOrderMap(2).toMap()
+	// Drop chunk values.
+	m = vs.ReadValue(vs.WriteValue(m).TargetRef()).(Map)
+	// Modify/query. Once upon a time this would crash.
+	fst, fstval := m.First()
+	m = m.Remove(fst)
+	assert.False(m.Has(fst))
+	{
+		fst, _ := m.First()
+		assert.True(m.Has(fst))
+	}
+	m = m.Set(fst, fstval)
+	assert.True(m.Has(fst))
+}
+
+func deriveMapHeight(m Map) uint64 {
+	// Note: not using mt.childRef.Height() because the purpose of this method is to be redundant.
+	height := uint64(1)
+	cm := m.(compoundMap)
+	if m2, ok := cm.getItem(0).(metaTuple).child.(compoundMap); ok {
+		height += deriveMapHeight(m2)
+	}
+	return height
 }
