+++ types/set_test.go
@@ -1,10 +1,164 @@
 package types
 
 import (
+	"math/rand"
+	"sort"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/suite"
 )
+
+const testSetSize = 5000
+
+type testSetGenFn func(v Number) Value
+type testSetLessFn func(x, y Value) bool
+
+type testSet struct {
+	values []Value
+	less   testSetLessFn
+	tr     *Type
+}
+
+func (ts testSet) Len() int {
+	return len(ts.values)
+}
+
+func (ts testSet) Less(i, j int) bool {
+	return ts.less(ts.values[i], ts.values[j])
+}
+
+func (ts testSet) Swap(i, j int) {
+	ts.values[i], ts.values[j] = ts.values[j], ts.values[i]
+}
+
+func (ts testSet) Remove(from, to int) testSet {
+	values := make([]Value, 0, len(ts.values)-(to-from))
+	values = append(values, ts.values[:from]...)
+	values = append(values, ts.values[to:]...)
+	return testSet{values, ts.less, ts.tr}
+}
+
+func (ts testSet) toSet() Set {
+	return NewTypedSet(ts.tr, ts.values...).(compoundSet)
+}
+
+func newTestSet(length int) testSet {
+	var values []Value
+	for i := 0; i < length; i++ {
+		values = append(values, Number(i))
+	}
+
+	return testSet{values,
+		func(x, y Value) bool {
+			return !y.(OrderedValue).Less(x.(OrderedValue))
+		},
+		MakeSetType(NumberType)}
+}
+
+func newTestSetWithGen(length int, gen testSetGenFn, less testSetLessFn, tr *Type) testSet {
+	s := rand.NewSource(4242)
+	used := map[int64]bool{}
+
+	var values []Value
+	for len(values) < length {
+		v := s.Int63() & 0xffffff
+		if _, ok := used[v]; !ok {
+			values = append(values, gen(Number(v)))
+			used[v] = true
+		}
+	}
+
+	return testSet{values, less, MakeSetType(tr)}
+}
+
+type setTestSuite struct {
+	collectionTestSuite
+	elems testSet
+}
+
+func newSetTestSuite(size uint, expectRefStr string, expectChunkCount int, expectPrependChunkDiff int, expectAppendChunkDiff int) *setTestSuite {
+	length := 1 << size
+	elems := newTestSet(length)
+	tr := MakeSetType(NumberType)
+	set := NewTypedSet(tr, elems.values...)
+	return &setTestSuite{
+		collectionTestSuite: collectionTestSuite{
+			col:                    set,
+			expectType:             tr,
+			expectLen:              uint64(length),
+			expectRef:              expectRefStr,
+			expectChunkCount:       expectChunkCount,
+			expectPrependChunkDiff: expectPrependChunkDiff,
+			expectAppendChunkDiff:  expectAppendChunkDiff,
+			validate: func(v2 Collection) bool {
+				l2 := v2.(Set)
+				out := []Value{}
+				l2.IterAll(func(v Value) {
+					out = append(out, v)
+				})
+				return valueSlicesEqual(elems.values, out)
+			},
+			prependOne: func() Collection {
+				dup := make([]Value, length+1)
+				dup[0] = Number(-1)
+				copy(dup[1:], elems.values)
+				return NewTypedSet(tr, dup...)
+			},
+			appendOne: func() Collection {
+				dup := make([]Value, length+1)
+				copy(dup, elems.values)
+				dup[len(dup)-1] = Number(length + 1)
+				return NewTypedSet(tr, dup...)
+			},
+		},
+		elems: elems,
+	}
+}
+
+func TestSetSuite1K(t *testing.T) {
+	suite.Run(t, newSetTestSuite(10, "sha1-53488a3b57c3db914bfc8d7863b91170602777d1", 16, 2, 2))
+}
+
+func TestSetSuite4K(t *testing.T) {
+	suite.Run(t, newSetTestSuite(12, "sha1-c1434d5fa03f233700b619ef39d679d3c38d811a", 3, 2, 2))
+}
+
+func getTestNativeOrderSet(scale int) testSet {
+	return newTestSetWithGen(int(setPattern)*scale, func(v Number) Value {
+		return v
+	}, func(x, y Value) bool {
+		return !y.(OrderedValue).Less(x.(OrderedValue))
+	}, NumberType)
+}
+
+func getTestRefValueOrderSet(scale int) testSet {
+	setType := MakeSetType(NumberType)
+	return newTestSetWithGen(int(setPattern)*scale, func(v Number) Value {
+		return NewTypedSet(setType, v)
+	}, func(x, y Value) bool {
+		return !y.Ref().Less(x.Ref())
+	}, setType)
+}
+
+func getTestRefToNativeOrderSet(scale int, vw ValueWriter) testSet {
+	refType := MakeRefType(NumberType)
+	return newTestSetWithGen(int(setPattern)*scale, func(v Number) Value {
+		return vw.WriteValue(v)
+	}, func(x, y Value) bool {
+		return !y.(Ref).TargetRef().Less(x.(Ref).TargetRef())
+	}, refType)
+}
+
+func getTestRefToValueOrderSet(scale int, vw ValueWriter) testSet {
+	setType := MakeSetType(NumberType)
+	refType := MakeRefType(setType)
+	return newTestSetWithGen(int(setPattern)*scale, func(v Number) Value {
+		return vw.WriteValue(NewTypedSet(setType, v))
+	}, func(x, y Value) bool {
+		return !y.(Ref).TargetRef().Less(x.(Ref).TargetRef())
+	}, refType)
+}
 
 func TestNewSet(t *testing.T) {
 	assert := assert.New(t)
@@ -107,6 +261,28 @@ func TestSetHas(t *testing.T) {
 	assert.False(s1.Has(Bool(false)))
 }
 
+func TestSetHas2(t *testing.T) {
+	if testing.Short() {
+		t.Skip("Skipping test in short mode.")
+	}
+	assert := assert.New(t)
+
+	vs := NewTestValueStore()
+	doTest := func(ts testSet) {
+		set := ts.toSet()
+		set2 := vs.ReadValue(vs.WriteValue(set).TargetRef()).(Set)
+		for _, v := range ts.values {
+			assert.True(set.Has(v))
+			assert.True(set2.Has(v))
+		}
+	}
+
+	doTest(getTestNativeOrderSet(16))
+	doTest(getTestRefValueOrderSet(2))
+	doTest(getTestRefToNativeOrderSet(2, vs))
+	doTest(getTestRefToValueOrderSet(2, vs))
+}
+
 func TestSetInsert(t *testing.T) {
 	assert := assert.New(t)
 	s := NewSet()
@@ -127,6 +303,46 @@ func TestSetInsert(t *testing.T) {
 	assert.True(s2.Has(v1))
 	assert.True(s2.Has(v2))
 	assert.True(s2.Has(v3))
+}
+
+func TestSetInsert2(t *testing.T) {
+	if testing.Short() {
+		t.Skip("Skipping test in short mode.")
+	}
+	assert := assert.New(t)
+
+	doTest := func(incr, offset int, ts testSet) {
+		expected := ts.toSet()
+		run := func(from, to int) {
+			actual := ts.Remove(from, to).toSet().Insert(ts.values[from:to]...)
+			assert.Equal(expected.Len(), actual.Len())
+			assert.True(expected.Equals(actual))
+		}
+		for i := 0; i < len(ts.values)-offset; i += incr {
+			run(i, i+offset)
+		}
+		run(len(ts.values)-offset, len(ts.values))
+		assert.Panics(func() {
+			expected.Insert(Bool(true))
+		}, "Should panic due to wrong type")
+	}
+
+	doTest(18, 3, getTestNativeOrderSet(9))
+	doTest(64, 1, getTestNativeOrderSet(32))
+	doTest(32, 1, getTestRefValueOrderSet(4))
+	doTest(32, 1, getTestRefToNativeOrderSet(4, NewTestValueStore()))
+	doTest(32, 1, getTestRefToValueOrderSet(4, NewTestValueStore()))
+}
+
+func TestSetInsertExistingValue(t *testing.T) {
+	assert := assert.New(t)
+
+	ts := getTestNativeOrderSet(2)
+	original := ts.toSet()
+	actual := original.Insert(ts.values[0])
+
+	assert.Equal(original.Len(), actual.Len())
+	assert.True(original.Equals(actual))
 }
 
 func TestSetRemove(t *testing.T) {
@@ -151,6 +367,107 @@ func TestSetRemove(t *testing.T) {
 	assert.True(s2.Has(v3))
 }
 
+func TestSetRemove2(t *testing.T) {
+	if testing.Short() {
+		t.Skip("Skipping test in short mode.")
+	}
+
+	assert := assert.New(t)
+
+	doTest := func(incr, offset int, ts testSet) {
+		whole := ts.toSet()
+		run := func(from, to int) {
+			expected := ts.Remove(from, to).toSet()
+			actual := whole.Remove(ts.values[from:to]...)
+			assert.Equal(expected.Len(), actual.Len())
+			assert.True(expected.Equals(actual))
+		}
+		for i := 0; i < len(ts.values)-offset; i += incr {
+			run(i, i+offset)
+		}
+		run(len(ts.values)-offset, len(ts.values))
+	}
+
+	doTest(18, 3, getTestNativeOrderSet(9))
+	doTest(64, 1, getTestNativeOrderSet(32))
+	doTest(32, 1, getTestRefValueOrderSet(4))
+	doTest(32, 1, getTestRefToNativeOrderSet(4, NewTestValueStore()))
+	doTest(32, 1, getTestRefToValueOrderSet(4, NewTestValueStore()))
+}
+
+func TestSetRemoveNonexistentValue(t *testing.T) {
+	assert := assert.New(t)
+
+	ts := getTestNativeOrderSet(2)
+	original := ts.toSet()
+	actual := original.Remove(Number(-1)) // rand.Int63 returns non-negative values.
+
+	assert.Equal(original.Len(), actual.Len())
+	assert.True(original.Equals(actual))
+}
+
+func TestSetUnion(t *testing.T) {
+	assert := assert.New(t)
+	assert.True(NewSet(Number(1), Number(2)).Union(
+		NewSet(Number(2), Number(3)),
+		NewSet(Number(-1)),
+		NewSet()).Equals(
+		NewSet(Number(1), Number(2), Number(3), Number(-1))))
+	assert.True(NewSet(Number(1)).Union().Equals(NewSet(Number(1))))
+}
+
+func TestSetUnion2(t *testing.T) {
+	if testing.Short() {
+		t.Skip("Skipping test in short mode.")
+	}
+	assert := assert.New(t)
+
+	doTest := func(ts testSet) {
+		cs := ts.toSet()
+		cs2 := cs.Union()
+		assert.True(cs.Equals(cs2))
+		cs3 := cs.Union(cs2)
+		assert.True(cs.Equals(cs3))
+		cs4 := cs.Union(cs2, cs3)
+		assert.True(cs.Equals(cs4))
+		emptySet := NewTypedSet(ts.tr)
+		cs5 := cs.Union(emptySet)
+		assert.True(cs.Equals(cs5))
+		cs6 := emptySet.Union(cs)
+		assert.True(cs.Equals(cs6))
+
+		r := rand.New(rand.NewSource(123))
+		subsetValues1 := make([]Value, 0, len(ts.values))
+		subsetValues2 := make([]Value, 0, len(ts.values))
+		subsetValues3 := make([]Value, 0, len(ts.values))
+		subsetValuesAll := make([]Value, 0, len(ts.values))
+		for _, v := range ts.values {
+			if r.Intn(3) == 0 {
+				subsetValues1 = append(subsetValues1, v)
+				subsetValuesAll = append(subsetValuesAll, v)
+			} else if r.Intn(3) == 0 {
+				subsetValues2 = append(subsetValues2, v)
+				subsetValuesAll = append(subsetValuesAll, v)
+			} else if r.Intn(3) == 0 {
+				subsetValues3 = append(subsetValues3, v)
+				subsetValuesAll = append(subsetValuesAll, v)
+			}
+		}
+
+		s1 := NewTypedSet(ts.tr, subsetValues1...)
+		s2 := NewTypedSet(ts.tr, subsetValues2...)
+		s3 := NewTypedSet(ts.tr, subsetValues3...)
+		sAll := NewTypedSet(ts.tr, subsetValuesAll...)
+
+		assert.True(s1.Union(s2, s3).Equals(sAll))
+	}
+
+	doTest(getTestNativeOrderSet(16))
+	doTest(getTestRefValueOrderSet(2))
+	doTest(getTestRefToNativeOrderSet(2, NewTestValueStore()))
+	doTest(getTestRefToValueOrderSet(2, NewTestValueStore()))
+}
+
 func TestSetFirst(t *testing.T) {
 	assert := assert.New(t)
 	s := NewSet()
@@ -163,24 +480,6 @@ func TestSetFirst(t *testing.T) {
 	assert.NotNil(s2.First())
 	s2 = s2.Remove(Number(2))
 	assert.Nil(s2.First())
-}
-
-func TestSetOfStruct(t *testing.T) {
-	assert := assert.New(t)
-
-	typ := MakeStructType("S1", TypeMap{
-		"o": NumberType,
-	})
-
-	elems := []Value{}
-	for i := 0; i < 200; i++ {
-		elems = append(elems, newStructFromData(structData{"o": Number(i)}, typ))
-	}
-
-	s := NewTypedSet(MakeSetType(typ), elems...)
-	for i := 0; i < 200; i++ {
-		assert.True(s.Has(elems[i]))
-	}
 }
 
 func TestSetIter(t *testing.T) {
@@ -202,6 +501,33 @@ func TestSetIter(t *testing.T) {
 	assert.True(acc.Empty())
 }
 
+func TestSetIter2(t *testing.T) {
+	assert := assert.New(t)
+
+	doTest := func(ts testSet) {
+		set := ts.toSet()
+		sort.Sort(ts)
+		idx := uint64(0)
+		endAt := uint64(setPattern)
+
+		set.Iter(func(v Value) (done bool) {
+			assert.True(ts.values[idx].Equals(v))
+			if idx == endAt {
+				done = true
+			}
+			idx++
+			return
+		})
+
+		assert.Equal(endAt, idx-1)
+	}
+
+	doTest(getTestNativeOrderSet(16))
+	doTest(getTestRefValueOrderSet(2))
+	doTest(getTestRefToNativeOrderSet(2, NewTestValueStore()))
+	doTest(getTestRefToValueOrderSet(2, NewTestValueStore()))
+}
+
 func TestSetIterAll(t *testing.T) {
 	assert := assert.New(t)
 	s := NewSet(Number(0), Number(1), Number(2), Number(3), Number(4))
@@ -212,6 +538,26 @@ func TestSetIterAll(t *testing.T) {
 		acc = acc.Insert(v)
 	})
 	assert.True(s.Equals(acc))
+}
+
+func TestSetIterAll2(t *testing.T) {
+	assert := assert.New(t)
+
+	doTest := func(ts testSet) {
+		set := ts.toSet()
+		sort.Sort(ts)
+		idx := uint64(0)
+
+		set.IterAll(func(v Value) {
+			assert.True(ts.values[idx].Equals(v))
+			idx++
+		})
+	}
+
+	doTest(getTestNativeOrderSet(16))
+	doTest(getTestRefValueOrderSet(2))
+	doTest(getTestRefToNativeOrderSet(2, NewTestValueStore()))
+	doTest(getTestRefToValueOrderSet(2, NewTestValueStore()))
 }
 
 func testSetOrder(assert *assert.Assertions, valueType *Type, value []Value, expectOrdering []Value) {
@@ -362,6 +708,32 @@ func TestSetFilter(t *testing.T) {
 	assert.True(NewSet(Number(0), Number(3)).Equals(s3))
 }
 
+func TestSetFilter2(t *testing.T) {
+	assert := assert.New(t)
+
+	doTest := func(ts testSet) {
+		set := ts.toSet()
+		sort.Sort(ts)
+		pivotPoint := 10
+		pivot := ts.values[pivotPoint]
+		actual := set.Filter(func(v Value) bool {
+			return ts.less(v, pivot)
+		})
+		assert.True(newTypedSet(ts.tr, ts.values[:pivotPoint+1]...).Equals(actual))
+
+		idx := 0
+		actual.IterAll(func(v Value) {
+			assert.True(ts.values[idx].Equals(v), "%v != %v", v, ts.values[idx])
+			idx++
+		})
+	}
+
+	doTest(getTestNativeOrderSet(16))
+	doTest(getTestRefValueOrderSet(2))
+	doTest(getTestRefToNativeOrderSet(2, NewTestValueStore()))
+	doTest(getTestRefToValueOrderSet(2, NewTestValueStore()))
+}
+
 func TestSetType(t *testing.T) {
 	assert := assert.New(t)
 
@@ -399,4 +771,77 @@ func TestSetChunks(t *testing.T) {
 	l2 := NewSet(NewTypedRefFromValue(Number(0)))
 	c2 := l2.Chunks()
 	assert.Len(c2, 1)
+}
+
+func TestSetChunks2(t *testing.T) {
+	assert := assert.New(t)
+
+	vs := NewTestValueStore()
+	doTest := func(ts testSet) {
+		set := ts.toSet()
+		set2chunks := vs.ReadValue(vs.WriteValue(set).TargetRef()).Chunks()
+		for i, r := range set.Chunks() {
+			assert.True(r.Type().Equals(set2chunks[i].Type()), "%s != %s", r.Type().Describe(), set2chunks[i].Type().Describe())
+		}
+	}
+
+	doTest(getTestNativeOrderSet(16))
+	doTest(getTestRefValueOrderSet(2))
+	doTest(getTestRefToNativeOrderSet(2, vs))
+	doTest(getTestRefToValueOrderSet(2, vs))
+}
+
+func TestSetFirstNNumbers(t *testing.T) {
+	assert := assert.New(t)
+
+	setType := MakeSetType(NumberType)
+
+	nums := generateNumbersAsValues(testSetSize)
+	s := NewTypedSet(setType, nums...)
+	assert.Equal("sha1-5937b476bb1d594e3a905c44f00863bd9ba1fb19", s.Ref().String())
+	height := deriveSetHeight(s)
+	cs := s.(compoundSet)
+	assert.Equal(height, cs.tuples[0].childRef.Height())
+}
+
+func TestSetRefOfStructFirstNNumbers(t *testing.T) {
+	if testing.Short() {
+		t.Skip("Skipping test in short mode.")
+	}
+	assert := assert.New(t)
+
+	structType, nums := generateNumbersAsStructs(testSetSize)
+	refOfTypeStructType := MakeRefType(structType)
+	setType := MakeSetType(refOfTypeStructType)
+	s := NewTypedSet(setType, nums...)
+	assert.Equal("sha1-3664c45fcbf64f1272956a7b93f27488ab0eb4f8", s.Ref().String())
+	height := deriveSetHeight(s)
+	cs := s.(compoundSet)
+	// height + 1 because the leaves are Ref values (with height 1).
+	assert.Equal(height+1, cs.tuples[0].childRef.Height())
+}
+
+func TestSetModifyAfterRead(t *testing.T) {
+	assert := assert.New(t)
+	vs := NewTestValueStore()
+	set := getTestNativeOrderSet(2).toSet()
+	// Drop chunk values.
+	set = vs.ReadValue(vs.WriteValue(set).TargetRef()).(Set)
+	// Modify/query. Once upon a time this would crash.
+	fst := set.First()
+	set = set.Remove(fst)
+	assert.False(set.Has(fst))
+	assert.True(set.Has(set.First()))
+	set = set.Insert(fst)
+	assert.True(set.Has(fst))
+}
+
+func deriveSetHeight(s Set) uint64 {
+	// Note: not using mt.childRef.Height() because the purpose of this method is to be redundant.
+	height := uint64(1)
+	cs := s.(compoundSet)
+	if s2, ok := cs.getItem(0).(metaTuple).child.(compoundSet); ok {
+		height += deriveSetHeight(s2)
+	}
+	return height
 }
