+++ types/util_test.go
@@ -2,10 +2,11 @@ package types
 
 import (
 	"github.com/attic-labs/noms/d"
+	"github.com/attic-labs/noms/ref"
 )
 
 var generateNumbersAsValues = func(n int) []Value {
-	d.Chk.True(n >= 0, "")
+	d.Chk.True(n > 0, "must be an integer greater than zero")
 	nums := []Value{}
 	for i := 0; i < n; i++ {
 		nums = append(nums, Number(i))
@@ -13,8 +14,18 @@ var generateNumbersAsValues = func(n int) []Value {
 	return nums
 }
 
+var generateNumbersAsStructs = func(n int) (*Type, []Value) {
+	d.Chk.True(n > 0, "must be an integer greater than zero")
+	structType := MakeStructType("num", TypeMap{"n": NumberType})
+	nums := []Value{}
+	for i := 0; i < n; i++ {
+		nums = append(nums, NewStruct(structType, structData{"n": Number(i)}))
+	}
+	return structType, nums
+}
+
 var generateNumbersAsRefOfStructs = func(n int) (*Type, []Value) {
-	d.Chk.True(n >= 0, "")
+	d.Chk.True(n > 0, "must be an integer greater than zero")
 	structType := MakeStructType("num", TypeMap{"n": NumberType})
 	vs := NewTestValueStore()
 	nums := []Value{}
@@ -25,12 +36,26 @@ var generateNumbersAsRefOfStructs = func(n int) (*Type, []Value) {
 	return structType, nums
 }
 
-var generateNumbersAsStructs = func(n int) (*Type, []Value) {
-	d.Chk.True(n >= 0, "")
-	structType := MakeStructType("num", TypeMap{"n": NumberType})
-	nums := []Value{}
-	for i := 0; i < n; i++ {
-		nums = append(nums, NewStruct(structType, structData{"n": Number(i)}))
+func chunkDiffCount(c1 []Ref, c2 []Ref) int {
+	count := 0
+	refs := make(map[ref.Ref]int)
+
+	for _, r := range c1 {
+		refs[r.TargetRef()]++
 	}
-	return structType, nums
+
+	for _, r := range c2 {
+		if c, ok := refs[r.TargetRef()]; ok {
+			if c == 1 {
+				delete(refs, r.TargetRef())
+			} else {
+				refs[r.TargetRef()] = c - 1
+			}
+		} else {
+			count++
+		}
+	}
+
+	count += len(refs)
+	return count
 }
